refactor: format.tsをモジュール構造に分割
